Move schema downloader tests to integration test target
