+++ Apollo.xcodeproj/xcshareddata/xcschemes/ApolloServerIntegrationTests.xcscheme
@@ -17,18 +17,6 @@
             default = "YES">
          </TestPlanReference>
       </TestPlans>
-      <Testables>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "DE6B15AB26152BE10068D642"
-               BuildableName = "ApolloServerIntegrationTests.xctest"
-               BlueprintName = "ApolloServerIntegrationTests"
-               ReferencedContainer = "container:Apollo.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
-      </Testables>
    </TestAction>
    <LaunchAction
       buildConfiguration = "Debug"
